clean example code
