+++ example/pubspec.yaml
@@ -30,7 +30,6 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  motion_toast: ^2.7.8
 
 dev_dependencies:
   flutter_test:
