+++ lib/src/web_hid_base.dart
@@ -22,7 +22,7 @@ class Hid extends Delegate<EventTarget> {
 
   void subscribeConnect(Function(HidDevice) listener) {
     _connectListener = allowInterop((event){
-      var e = getObjectProperty(event, 'device');
+      var e = getPropertyFrom(event, 'device');
       listener(HidDevice._(e));
     });
     setProperty('onconnect', _connectListener);
@@ -35,7 +35,7 @@ class Hid extends Delegate<EventTarget> {
 
   void subscribeDisconnect(Function(HidDevice) listener) {
     _disconnectListener = allowInterop((event){
-      var e = getObjectProperty(event, 'device');
+      var e = getPropertyFrom(event, 'device');
       listener(HidDevice._(e));
     });
     setProperty('ondisconnect', _disconnectListener);
@@ -64,4 +64,48 @@ class RequestOptionsFilter {
     int usage,
     int usagePage,
   });
+}
+
+class HidDevice extends Delegate<EventTarget> {
+  HidDevice._(EventTarget delegate) : super(delegate);
+  late EventListener _eventListener;
+
+  Future<void> open() {
+    var promise = callMethod('open');
+    return promiseToFuture(promise);
+  }
+
+  Future<void> close() {
+    var promise = callMethod('close');
+    return promiseToFuture(promise);
+  }
+
+  bool get opened => getProperty('opened');
+
+  Future<void> sendReport(int requestId, TypedData data) {
+    var promise = callMethod('sendReport', [requestId, data]);
+    return promiseToFuture(promise);
+  }
+
+  void subscribeInputReport(Function(Uint8List) listener) {
+    _eventListener = allowInterop((Event event) {
+      ByteData blockData = getPropertyFrom(event, 'data');
+      var reportId = getPropertyFrom(event, 'reportId');
+
+      List<int> dataList = [reportId];
+      dataList.addAll(blockData.buffer.asUint8List());
+      listener(Uint8List.fromList(dataList));
+    });
+    setProperty('oninputreport', _eventListener);
+  }
+
+  void unsubscribeInputReport(Function() listener) {
+    setProperty('oninputreport', null);
+    listener();
+  }
+
+  Future<void> sendFeatureReport(int requestId, TypedData data) {
+    var promise = callMethod('sendFeatureReport', [requestId, data]);
+    return promiseToFuture(promise);
+  }
 }
